Refactor+Design: Make room for a homescreen and easy scene switching
