+++ src/CMakeLists.txt
@@ -11,12 +11,13 @@ project(
 set(CMAKE_CXX_FLAGS)
 
 add_library(BotsenCore SHARED
-    ${CMAKE_CURRENT_SOURCE_DIR}/object.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/raylibCircleExt.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/collisions.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/Object.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/RaylibCircleExt.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/Collisions.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/Scenes.cpp
     )
 target_include_directories(BotsenCore PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
-add_executable(Botsen ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp)
+add_executable(Botsen ${CMAKE_CURRENT_SOURCE_DIR}/Main.cpp)
 target_link_libraries(Botsen BotsenCore raylib)
