+++ src/Collisions.cpp
@@ -1,4 +1,4 @@
-#include <collisions.h>
+#include <Collisions.h>
 
 namespace Collision {
