+++ src/RaylibCircleExt.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <raylib.h>
-#include <object.h>
+#include <Object.h>
 
 namespace RaylibExt {
